sechster commit

+++ app/src/main/java/com/example/myreptil/data/datamodels/Tier.kt
@@ -25,7 +25,7 @@ data class Tier(
     var Generation: String,
     val LetzteHaeutung: Double,
     val LetzteFuetterung: Double,
-    val FuetterungsIntervall: Int,
+    val FuetterungsIntervall: Double,
     val Tierarzt: String,
     val LetzterArztbesuch: Double,
     val LetztesEreignis: String,
